자기정보수정화면 에러수정

+++ src/main/kotlin/co/brainz/itsm/user/service/UserService.kt
@@ -81,7 +81,7 @@ class UserService(private val certificationRepository: CertificationRepository,
     /**
      * 사용자의 KEY로 정보를 수정한다.
      */
-    fun updateUserEdit(userUpdateDto: UserUpdateDto): String {
+    fun updateUserEdit(userUpdateDto: UserUpdateDto, target: String): String {
         var code: String = userEditValid(userUpdateDto)
         when (code) {
             UserConstants.UserEditStatus.STATUS_VALID_SUCCESS.code -> {
@@ -100,12 +100,14 @@ class UserService(private val certificationRepository: CertificationRepository,
                 logger.debug("targetEntity {}, update {}", targetEntity, userUpdateDto)
                 userRepository.save(targetEntity)
 
-                userEntity.userRoleMapEntities.forEach {
-                    userRoleMapRepository.deleteById(AliceUserRoleMapPk(userUpdateDto.userKey, it.role.roleId))
-                }
+                if (target == "userEdit") {
+                    userEntity.userRoleMapEntities.forEach {
+                        userRoleMapRepository.deleteById(AliceUserRoleMapPk(userUpdateDto.userKey, it.role.roleId))
+                    }
 
-                userUpdateDto.roles!!.forEach {
-                    userRoleMapRepository.save(AliceUserRoleMapEntity(targetEntity, roleRepository.findByRoleId(it)))
+                    userUpdateDto.roles!!.forEach {
+                        userRoleMapRepository.save(AliceUserRoleMapEntity(targetEntity, roleRepository.findByRoleId(it)))
+                    }
                 }
 
                 code =  if (targetEntity.email ==  emailConfirmVal && userUpdateDto.userKey == aliceUserDto.userKey) {
